Worker: Refactor flows for crawl one and two users
